perf(prometheus): avoid unused diagnostic log work
Exclude log.record from the metric exporter subscription so enabling Prometheus does not activate diagnostic log preparation, redaction, and cloning. Preserve all metric events, private-content policy, and independent log subscribers. Move the existing bus regression into a focused subscription test and cover ignored log payloads and subscriber coexistence.
