+++ extensions/diagnostics-prometheus/src/service.subscription.test.ts
@@ -1,0 +1,73 @@
+import {
+  emitDiagnosticEvent,
+  emitTrustedDiagnosticEventWithPrivateData,
+  waitForDiagnosticEventsDrained,
+} from "openclaw/plugin-sdk/diagnostic-runtime";
+import { onTrustedInternalDiagnosticEvent } from "openclaw/plugin-sdk/plugin-test-runtime";
+import { expect, it, vi } from "vitest";
+import { createDiagnosticsPrometheusExporter } from "./service.js";
+
+it("records terminal metrics without reading logs or private diagnostic content", async () => {
+  const exporter = createDiagnosticsPrometheusExporter();
+  const readPrivateContent = vi.fn(() => ({ toolInput: { text: "private tool input" } }));
+  const readLogAttribute = vi.fn(() => "synthetic log attribute");
+  const emitLog = () =>
+    emitDiagnosticEvent({
+      type: "log.record",
+      level: "INFO",
+      message: "synthetic log",
+      attributes: {
+        get detail() {
+          return readLogAttribute();
+        },
+      },
+    });
+  exporter.service.start({
+    config: {},
+    stateDir: "/tmp/openclaw-prometheus-test",
+    logger: { info: vi.fn(), warn: vi.fn(), error: vi.fn(), debug: vi.fn() },
+    internalDiagnostics: {
+      emit() {},
+      onEvent: onTrustedInternalDiagnosticEvent,
+    },
+  });
+  try {
+    emitLog();
+    emitTrustedDiagnosticEventWithPrivateData(
+      {
+        type: "tool.execution.completed",
+        runId: "run-1",
+        toolCallId: "call-1",
+        toolName: "synthetic",
+        toolSource: "core",
+        durationMs: 3,
+      },
+      {
+        get toolContent() {
+          return readPrivateContent();
+        },
+      },
+    );
+    await waitForDiagnosticEventsDrained();
+    expect(exporter.render()).toContain(
+      'openclaw_tool_execution_total{error_category="none",outcome="completed",params_kind="unknown",tool="synthetic",tool_owner="none",tool_source="core"} 1',
+    );
+    expect(readPrivateContent).not.toHaveBeenCalled();
+    expect(readLogAttribute).not.toHaveBeenCalled();
+    const logs = vi.fn();
+    const stopLogs = onTrustedInternalDiagnosticEvent(logs, { include: ["log.record"] });
+    try {
+      emitLog();
+      await waitForDiagnosticEventsDrained();
+      expect(logs).toHaveBeenCalledOnce();
+      expect(logs.mock.calls[0]?.[0].attributes).toEqual({ detail: "synthetic log attribute" });
+      expect(readLogAttribute).toHaveBeenCalledOnce();
+    } finally {
+      stopLogs();
+    }
+  } finally {
+    exporter.service.stop?.();
+    await waitForDiagnosticEventsDrained();
+  }
+  expect(exporter.render()).toBe("");
+});
